refactor raytracing renderer

+++ src/Application.cpp
@@ -242,9 +242,9 @@ void NextRendererApplication<Renderer>::Render(VkCommandBuffer commandBuffer, co
     stats.CamPosY = modelViewController_.Position()[1];
     stats.CamPosZ = modelViewController_.Position()[2];
 
-    stats.InstanceCount = scene_->Nodes().size();
+    stats.InstanceCount = static_cast<uint32_t>(scene_->Nodes().size());
     stats.TriCount = scene_->GetIndicesCount() / 3;
-    stats.TextureCount = scene_->TextureSamplers().size();
+    stats.TextureCount = static_cast<uint32_t>(scene_->TextureSamplers().size());
     stats.ComputePassCount = 0;
 
     if (userSettings_.IsRayTraced)
